+++ modules/Scheduling/MassDrops.php
@@ -224,7 +224,7 @@ if ( ! $_REQUEST['modfunc'] )
 	}
 
 	$extra['link'] = [ 'FULL_NAME' => false ];
-	$extra['SELECT'] = ",CAST (NULL AS CHAR(1)) AS CHECKBOX";
+	$extra['SELECT'] = ",NULL AS CHECKBOX";
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
 	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( '', 'STUDENT_ID', 'student' ) ];
 	$extra['new'] = true;