+++ modules/Scheduling/MassSchedule.php
@@ -187,7 +187,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 	$extra['link'] = [ 'FULL_NAME' => false ];
 
-	$extra['SELECT'] = ",CAST (NULL AS CHAR(1)) AS CHECKBOX";
+	$extra['SELECT'] = ",NULL AS CHECKBOX";
 
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];