+++ modules/Students/AssignOtherInfo.php
@@ -202,7 +202,7 @@ echo ErrorMessage( $warning, 'warning' );
 if ( ! $_REQUEST['modfunc'] )
 {
 	$extra['link'] = [ 'FULL_NAME' => false ];
-	$extra['SELECT'] = ",CAST (NULL AS CHAR(1)) AS CHECKBOX";
+	$extra['SELECT'] = ",NULL AS CHECKBOX";
 
 	if ( $_REQUEST['search_modfunc'] === 'list' )
 	{